cambios de move
